remove unused stuff, further separate prototype bits
