+++ backend/apps/readings/views.py
@@ -6,24 +6,14 @@ from rest_framework import generics
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 
-from .models import StoryPrototype, Student, Document, StudentReadingData
+from .models import Student, Document, StudentReadingData
 from .analysis import RereadingAnalysis
 from .serializers import (
-    StoryPrototypeSerializer, StudentPrototypeSerializer, AnalysisSerializer,
-    DocumentSerializer, StudentSerializer, DocumentAnalysisSerializer,
-    StudentReadingDataSerializer, ReadingSerializer)
-
-
-class ListDocument(generics.ListCreateAPIView):
-    """View a list of documents or create a new one"""
-    queryset = Document.objects.all()
-    serializer_class = DocumentSerializer
-
-
-class DetailDocument(generics.RetrieveUpdateDestroyAPIView):
-    """Get a single document or update/delete it"""
-    queryset = Document.objects.all()
-    serializer_class = DocumentSerializer
+    DocumentSerializer,
+    ReadingSerializer,
+    StudentReadingDataSerializer,
+    StudentSerializer,
+)
 
 
 class Reading:
@@ -50,17 +40,6 @@ def reading_view(request, pk):
     return Response(serializer.data)
 
 
-@api_view(['GET'])
-def document_analysis(request):
-    """
-    Init a DocumentAnalysis, and serialize it to send to the frontend.
-    We are hardcoding the recitatif document for prototyping the DocumentAnalysis view
-    """
-    queryset = Document.objects.get(pk=1)
-    serializer = DocumentAnalysisSerializer(queryset)
-    return Response(serializer.data)
-
-
 class ListStudent(generics.ListCreateAPIView):
     """View a list of students or create a new one"""
     queryset = Student.objects.all()
@@ -83,33 +62,6 @@ class ListReadingData(generics.ListCreateAPIView):
     """View all instances of reading data"""
     queryset = StudentReadingData.objects.all()
     serializer_class = StudentReadingDataSerializer
-
-################################################################################
-# Prototype views
-# The API endpoints below were for the prototype from summer 2019
-################################################################################
-class ListStudentPrototype(generics.ListCreateAPIView):
-    """ View a list of students of create a new one """
-    queryset = Student.objects.all()
-    serializer_class = StudentPrototypeSerializer
-
-
-class DetailStudentPrototype(generics.RetrieveUpdateDestroyAPIView):
-    """ Get a single Student or update/delete it """
-    queryset = Student.objects.all()
-    serializer_class = StudentPrototypeSerializer
-
-
-class ListStoryPrototype(generics.ListCreateAPIView):
-    """ Get a list of story objects """
-    queryset = StoryPrototype.objects.all()
-    serializer_class = StoryPrototypeSerializer
-
-
-class DetailStoryPrototype(generics.RetrieveUpdateDestroyAPIView):
-    """ Get a single story object, or update/delete it """
-    queryset = StoryPrototype.objects.all()
-    serializer_class = StoryPrototypeSerializer
 
 
 @api_view(['GET'])
