+++ frontend/src/prototype/student_view.css
